% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_convex_hull_volume.R
\name{calculate_convex_hull_volume}
\alias{calculate_convex_hull_volume}
\title{Calculate Convex Hull Volume}
\usage{
calculate_convex_hull_volume(points)
}
\arguments{
\item{points}{A data frame or matrix with columns 'X', 'Y', and 'Z' representing point coordinates.}
}
\value{
Numeric value representing the volume of the convex hull.
}
\description{
Calculates the volume of a convex hull formed by a set of 3D points.
}
