% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_marker_data.R
\name{process_marker_data}
\alias{process_marker_data}
\title{Deprecated: Process Marker Data}
\usage{
process_marker_data(data, convert_to_cm = FALSE)
}
\arguments{
\item{data}{A data frame containing marker coordinate data for all time frames.}

\item{convert_to_cm}{Logical, if TRUE, converts units from mm to cm.}
}
\value{
A data frame of the same format as reformat_marker_data().
}
\description{
This function is deprecated. Please use \code{\link{reformat_marker_data}} instead.
}
\examples{
data("sample_data")
# Using the old function will show a warning and call reformat_marker_data()
processed_data <- process_marker_data(head(sample_data), convert_to_cm = TRUE)
head(processed_data)

}
