% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_position.R
\name{adj_position}
\alias{adj_position}
\title{Adjust Marker Positions Towards Center}
\usage{
adj_position(marker_coords, center_coords, distance = 1)
}
\arguments{
\item{marker_coords}{Numeric vector of length 3 representing the marker coordinates (X, Y, Z).}

\item{center_coords}{Numeric vector of length 3 representing the center coordinates (X, Y, Z).}

\item{distance}{Numeric value indicating the distance to adjust towards the center (default is 1 cm).}
}
\value{
A numeric vector of length 3 representing the adjusted marker coordinates.
}
\description{
Adjusts the position of markers by moving them towards the center by a specified distance.
}
